+++ src/main/java/com/carumuch/capstone/bidding/application/BidService.java
@@ -7,9 +7,9 @@ import com.carumuch.capstone.bidding.presentation.dto.BidPageResDto;
 import com.carumuch.capstone.bidding.presentation.dto.BidResDto;
 import com.carumuch.capstone.bidding.domain.BidRepository;
 import com.carumuch.capstone.common.legacy.exception.CustomException;
-import com.carumuch.capstone.estimate.model.Estimate;
-import com.carumuch.capstone.estimate.model.type.EstimateStatus;
-import com.carumuch.capstone.estimate.repository.EstimateRepository;
+import com.carumuch.capstone.estimate.domain.Estimate;
+import com.carumuch.capstone.estimate.domain.EstimateStatus;
+import com.carumuch.capstone.estimate.domain.EstimateRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
