+++ src/main/java/com/carumuch/capstone/bidding/domain/Bid.java
@@ -2,7 +2,7 @@ package com.carumuch.capstone.bidding.domain;
 
 import com.carumuch.capstone.bodyshop.domain.BodyShop;
 import com.carumuch.capstone.common.legacy.base.BaseCreateByEntity;
-import com.carumuch.capstone.estimate.model.Estimate;
+import com.carumuch.capstone.estimate.domain.Estimate;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
