+++ src/main/java/com/carumuch/capstone/bodyshop/application/BodyShopService.java
@@ -17,12 +17,12 @@ import com.carumuch.capstone.common.legacy.exception.ErrorCode;
 import com.carumuch.capstone.common.legacy.exception.CustomException;
 import com.carumuch.capstone.user.domain.User;
 import com.carumuch.capstone.user.domain.UserLegacyRepository;
-import com.carumuch.capstone.estimate.model.Estimate;
-import com.carumuch.capstone.estimate.dto.EstimateDetailResDto;
-import com.carumuch.capstone.estimate.dto.EstimateSearchReqDto;
-import com.carumuch.capstone.estimate.dto.EstimateSearchResDto;
-import com.carumuch.capstone.estimate.repository.EstimateRepository;
-import com.carumuch.capstone.estimate.repository.custom.EstimateRepositoryCustom;
+import com.carumuch.capstone.estimate.domain.Estimate;
+import com.carumuch.capstone.estimate.presentation.dto.EstimateDetailResDto;
+import com.carumuch.capstone.estimate.presentation.dto.EstimateSearchReqDto;
+import com.carumuch.capstone.estimate.presentation.dto.EstimateSearchResDto;
+import com.carumuch.capstone.estimate.domain.EstimateRepository;
+import com.carumuch.capstone.estimate.domain.EstimateRepositoryCustom;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
