+++ src/main/java/com/carumuch/capstone/estimate/application/EstimateService.java
@@ -1,22 +1,22 @@
-package com.carumuch.capstone.estimate.service;
+package com.carumuch.capstone.estimate.application;
 
 import com.carumuch.capstone.bidding.domain.BidStatus;
 import com.carumuch.capstone.bidding.domain.BidRepository;
-import com.carumuch.capstone.estimate.dto.EstimateAIRepairCostReqDto;
-import com.carumuch.capstone.estimate.dto.EstimateByVehiclePageResDto;
-import com.carumuch.capstone.estimate.dto.EstimateDetailResDto;
-import com.carumuch.capstone.estimate.dto.EstimatePageResDto;
-import com.carumuch.capstone.estimate.dto.EstimateRegistrationReqDto;
-import com.carumuch.capstone.estimate.dto.EstimateStatusUpdateReqDto;
-import com.carumuch.capstone.estimate.dto.EstimateUpdateReqDto;
-import com.carumuch.capstone.estimate.repository.EstimateRepository;
+import com.carumuch.capstone.estimate.presentation.dto.EstimateAIRepairCostReqDto;
+import com.carumuch.capstone.estimate.presentation.dto.EstimateByVehiclePageResDto;
+import com.carumuch.capstone.estimate.presentation.dto.EstimateDetailResDto;
+import com.carumuch.capstone.estimate.presentation.dto.EstimatePageResDto;
+import com.carumuch.capstone.estimate.presentation.dto.EstimateRegistrationReqDto;
+import com.carumuch.capstone.estimate.presentation.dto.EstimateStatusUpdateReqDto;
+import com.carumuch.capstone.estimate.presentation.dto.EstimateUpdateReqDto;
+import com.carumuch.capstone.estimate.domain.EstimateRepository;
 import com.carumuch.capstone.common.legacy.exception.ErrorCode;
 import com.carumuch.capstone.common.legacy.exception.CustomException;
 import com.carumuch.capstone.user.domain.User;
 import com.carumuch.capstone.user.domain.UserLegacyRepository;
-import com.carumuch.capstone.estimate.model.Estimate;
+import com.carumuch.capstone.estimate.domain.Estimate;
 import com.carumuch.capstone.vehicle.model.Vehicle;
-import com.carumuch.capstone.estimate.model.type.EstimateStatus;
+import com.carumuch.capstone.estimate.domain.EstimateStatus;
 import com.carumuch.capstone.vehicle.repository.VehicleRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
