+++ src/main/java/com/carumuch/capstone/estimate/presentation/dto/EstimateAnalysisReqDto.java
@@ -1,4 +1,4 @@
-package com.carumuch.capstone.estimate.dto;
+package com.carumuch.capstone.estimate.presentation.dto;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
