+++ src/main/java/com/carumuch/capstone/estimate/presentation/dto/EstimateDetailResDto.java
@@ -1,6 +1,6 @@
-package com.carumuch.capstone.estimate.dto;
+package com.carumuch.capstone.estimate.presentation.dto;
 
-import com.carumuch.capstone.estimate.model.Estimate;
+import com.carumuch.capstone.estimate.domain.Estimate;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Builder;
 import lombok.Getter;
